Renamed project to Hades, so it fits better with new vision.
Started on improving readme.

Added build action. Not sure if it works tho.
