+++ .github/workflows/test.yaml
@@ -1,4 +1,4 @@
-name: Run Tests
+name: Run tests
 
 on:
   push:
