+++ .gitignore
@@ -21,4 +21,4 @@ saves/
 aseprite/
 
 # Temp files
-game/TEMP-FILES/
+hades/TEMP-FILES/
