+++ README.md
@@ -1,1 +1,7 @@
+# Hades
+
+![GitHub](https://img.shields.io/github/license/Aspect1103/Hades)
+
+[![Language grade: Python](https://img.shields.io/lgtm/grade/python/g/Aspect1103/Hades.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/Aspect1103/Hades/context:python)
+
 A procedurally generated dungeon crawler game made in Python using Arcade.
