+++ hades/common.py
