+++ hades/constants/__init__.py
