+++ hades/constants/game_object.py
