+++ hades/constants/general.py
