+++ hades/constants/generation.py
