+++ hades/game_objects/base.py
@@ -10,29 +10,29 @@ from typing import TYPE_CHECKING
 import arcade
 
 # Custom
-from game.constants.game_object import (
+from hades.constants.game_object import (
     ARMOUR_REGEN_AMOUNT,
     ARMOUR_REGEN_WAIT,
     SPRITE_SCALE,
     EntityAttributeType,
     ObjectID,
 )
-from game.game_objects.attack import create_attack
-from game.textures import grid_pos_to_pixel
+from hades.game_objects.attack import create_attack
+from hades.textures import grid_pos_to_pixel
 
 if TYPE_CHECKING:
-    from game.constants.game_object import (
+    from hades.constants.game_object import (
         AttackAlgorithmType,
         AttackData,
         BaseData,
         EntityAttributeData,
         EntityData,
     )
-    from game.game_objects.attack import AttackBase
-    from game.game_objects.attribute import EntityAttribute
-    from game.game_objects.player import Player
-    from game.physics import PhysicsEngine
-    from game.views.game_view import Game
+    from hades.game_objects.attack import AttackBase
+    from hades.game_objects.attribute import EntityAttribute
+    from hades.game_objects.player import Player
+    from hades.physics import PhysicsEngine
+    from hades.views.game_view import Game
 
 __all__ = (
     "CollectibleTile",
