+++ hades/resources/shader scripts/melee_geometry.glsl
