+++ hades/resources/shader scripts/melee_vertex.glsl
