+++ hades/resources/textures/armour_boost_potion.png
