+++ hades/resources/textures/armour_potion.png
