+++ hades/resources/textures/enemy_idle.png
