+++ hades/resources/textures/fire_rate_boost_potion.png
