+++ hades/resources/textures/health_boost_potion.png
