+++ hades/resources/textures/health_potion.png
