+++ hades/resources/textures/player_idle.png
