+++ hades/resources/textures/speed_boost_potion.png
