+++ hades/resources/textures/wall.png
