+++ hades/textures.py
@@ -9,7 +9,7 @@ import pathlib
 import arcade
 
 # Custom
-from game.constants.game_object import SPRITE_SIZE
+from hades.constants.game_object import SPRITE_SIZE
 
 __all__ = (
     "grid_pos_to_pixel",
