+++ tests/generation/test_primitives.py
@@ -6,8 +6,8 @@ import numpy as np
 import pytest
 
 # Custom
-from game.constants.generation import TileType
-from game.generation.primitives import Point, Rect
+from hades.constants.generation import TileType
+from hades.generation.primitives import Point, Rect
 
 __all__ = ()
